dp_publish in rh8

+++ plugin/app/plugins/pipeline_plugin/profiles/package/redhat/createrepo.sh
@@ -3,6 +3,8 @@ function createrepo(){
     local directory=$1
     [[ "$1" == "" ]] && _log "[ERROR] directory parameter is mandatory" && exit 1
     eval "$REMOTE_COMMAND mkdir -p $directory" && \
-       eval "$REMOTE_COMMAND /usr/bin/createrepo -s sha -d --update $directory" && \
-       eval "$REMOTE_COMMAND 'rm -rf $directory/repoview;/usr/bin/repoview $directory'"
+      eval "$REMOTE_COMMAND /usr/bin/createrepo -s sha -d --update $directory"
+    [[ $? != 0 ]] && return 1
+    which repoview 2>/dev/null && \
+      eval "$REMOTE_COMMAND 'rm -rf $directory/repoview;/usr/bin/repoview $directory'"
 }
